ActorState: move files to a dedicated actor_state directory
